Pre-tested version

+++ bin/gencatalog.sh
@@ -11,17 +11,34 @@ if [ ! -d $releasedir ] ; then
 fi
 
 # Clean the release directory
-( cd $releasedir; rm -rf * )
+# Clean release directory before start.
+rm -rf $releasedir/*
 
-echo "Exporting books..."
+echo -n "Exporting books..."
 mysqldump -h $host -u $user -p$password ekkitab_books books > $releasedir/books.sql
-echo "Copying search related files.."
+echo "done."
+echo -n "Zipping books data..."
+( cd $releasedir ; zip -q books.zip books.sql && rm books.sql )
+echo "done."
+if [ ! -f $releasedir/books.zip ] ; then
+    echo "FATAL: Zip process appears to have failed. No books.zip file found."
+    exit 1
+fi
+
+echo -n "Copying search related files.."
 cp -r $EKKITAB_HOME/magento/search_index_dir $releasedir
 cp -r $EKKITAB_HOME/magento/search_index_dir_spell_author $releasedir
 cp -r $EKKITAB_HOME/magento/search_index_dir_spell_title $releasedir
 cp -r $EKKITAB_HOME/magento/categories.xml $releasedir
-echo "Creating release date file"
+echo "done."
+
+echo -n "Copying synch file to synchronize release on server..."
+cp $EKKITAB_HOME/bin/synchcatalog.sh  $releasedir
+echo "done."
+
+echo -n "Creating release date file..."
 echo `date +%D' 'at' '%T` > $releasedir/releasedate
+echo "done."
 echo "Release directory created on `date +%D' 'at' '%T`"
